pyproject.toml
[project]
name = "optimade-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
]
